# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MyHUB Internet Café Management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tables
    description: Table management
  - name: products
    description: Menu product management
  - name: orders
    description: Order management
  - name: reservations
    description: Reservation management
  - name: admin
    description: Admin authentication
  - name: dashboard
    description: Dashboard summaries

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tables:
    get:
      operationId: getTables
      tags: [tables]
      summary: List all tables
      responses:
        "200":
          description: List of tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create a new table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTableBody"
      responses:
        "201":
          description: Created table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"

  /tables/{id}:
    get:
      operationId: getTable
      tags: [tables]
      summary: Get a single table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Table details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    put:
      operationId: updateTable
      tags: [tables]
      summary: Update a table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTableBody"
      responses:
        "200":
          description: Updated table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    delete:
      operationId: deleteTable
      tags: [tables]
      summary: Delete a table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: tableId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of orders with items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithItems"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"

  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"

  /orders/{id}/payment:
    put:
      operationId: updateOrderPayment
      tags: [orders]
      summary: Update order payment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderPaymentBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"

  /reservations:
    get:
      operationId: getReservations
      tags: [reservations]
      summary: List all reservations
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a reservation (no account needed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationBody"
      responses:
        "201":
          description: Created reservation with code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"

  /reservations/{code}:
    get:
      operationId: getReservationByCode
      tags: [reservations]
      summary: Get reservation by code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reservation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"

  /reservations/{id}/status:
    put:
      operationId: updateReservationStatus
      tags: [reservations]
      summary: Update reservation status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReservationStatusBody"
      responses:
        "200":
          description: Updated reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard overview stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/recent-orders:
    get:
      operationId: getRecentOrders
      tags: [dashboard]
      summary: Get recent orders for activity feed
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithItems"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    Table:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [available, occupied]
        capacity:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - status
        - capacity
        - createdAt

    CreateTableBody:
      type: object
      properties:
        name:
          type: string
        capacity:
          type: integer
      required:
        - name
        - capacity

    UpdateTableBody:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [available, occupied]
        capacity:
          type: integer

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
        category:
          type: string
        description:
          type: string
        available:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - price
        - category
        - available
        - createdAt

    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        category:
          type: string
        description:
          type: string
        available:
          type: boolean
      required:
        - name
        - price
        - category

    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        category:
          type: string
        description:
          type: string
        available:
          type: boolean

    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        productName:
          type: string
        productPrice:
          type: number
      required:
        - id
        - orderId
        - productId
        - quantity
        - productName
        - productPrice

    OrderWithItems:
      type: object
      properties:
        id:
          type: integer
        tableId:
          type: integer
        tableName:
          type: string
        status:
          type: string
          enum: [pending, preparing, completed]
        paymentStatus:
          type: string
          enum: [unpaid, paid]
        totalAmount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
      required:
        - id
        - tableId
        - tableName
        - status
        - paymentStatus
        - totalAmount
        - items
        - createdAt

    CreateOrderBody:
      type: object
      properties:
        tableId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
            required:
              - productId
              - quantity
      required:
        - tableId
        - items

    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, preparing, completed]
      required:
        - status

    UpdateOrderPaymentBody:
      type: object
      properties:
        paymentStatus:
          type: string
          enum: [unpaid, paid]
      required:
        - paymentStatus

    Reservation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        dateTime:
          type: string
        code:
          type: string
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        partySize:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - dateTime
        - code
        - status
        - partySize
        - createdAt

    CreateReservationBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        dateTime:
          type: string
        partySize:
          type: integer
      required:
        - name
        - phone
        - dateTime
        - partySize

    UpdateReservationStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled]
      required:
        - status

    AdminLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        admin:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
          required:
            - id
            - username
      required:
        - success

    AdminSession:
      type: object
      properties:
        authenticated:
          type: boolean
        admin:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
          required:
            - id
            - username
      required:
        - authenticated

    DashboardSummary:
      type: object
      properties:
        totalTables:
          type: integer
        occupiedTables:
          type: integer
        availableTables:
          type: integer
        totalOrders:
          type: integer
        pendingOrders:
          type: integer
        preparingOrders:
          type: integer
        completedOrders:
          type: integer
        todayRevenue:
          type: number
        unpaidAmount:
          type: number
        totalReservations:
          type: integer
        pendingReservations:
          type: integer
        confirmedReservations:
          type: integer
      required:
        - totalTables
        - occupiedTables
        - availableTables
        - totalOrders
        - pendingOrders
        - preparingOrders
        - completedOrders
        - todayRevenue
        - unpaidAmount
        - totalReservations
        - pendingReservations
        - confirmedReservations